Move CustomDist code to its own module
