+++ pymc/distributions/__init__.py
@@ -50,6 +50,7 @@ from pymc.distributions.continuous import (
     Wald,
     Weibull,
 )
+from pymc.distributions.custom import CustomDist, DensityDist
 from pymc.distributions.discrete import (
     Bernoulli,
     BetaBinomial,
@@ -66,8 +67,6 @@ from pymc.distributions.discrete import (
 )
 from pymc.distributions.distribution import (
     Continuous,
-    CustomDist,
-    DensityDist,
     DiracDelta,
     Discrete,
     Distribution,
